requirements.txt
Flask
pika
celery
celery[redis]
psycopg2-binary
python-dotenv
requests
gunicorn